// src/app/shared/badge/badge.types.ts
export type BadgeVariant = 'soft' | 'solid' | 'outline' | 'dot' | 'ghost';

export type BadgeStatus =
  | 'primary'
  | 'secondary'
  | 'success'
  | 'warning'
  | 'danger'
  | 'info'
  | 'neutral'
  | 'pending'
  | 'active'
  | 'inactive'
  | 'new'
  | 'completed';

export type BadgeSize = 'sm' | 'md' | 'lg';